🐞 fix: issued job can not be updated

+++ src/modules/ordered-service/command/delete-ordered-service/delete-ordered-service.service.ts
@@ -6,6 +6,8 @@ import { DeleteOrderedServiceCommand } from './delete-ordered-service.command'
 import { OrderedServiceRepositoryPort } from '../../database/ordered-service.repository.port'
 import { ORDERED_SERVICE_REPOSITORY } from '../../ordered-service.di-token'
 import { OrderedServiceNotFoundException } from '../../domain/ordered-service.error'
+import { IssuedJobUpdateException, JobNotFoundException } from '../../../ordered-job/domain/job.error'
+import { InvoiceNotFoundException } from '../../../invoice/domain/invoice.error'
 
 @CommandHandler(DeleteOrderedServiceCommand)
 export class DeleteOrderedServiceService implements ICommandHandler {
@@ -19,6 +21,18 @@ export class DeleteOrderedServiceService implements ICommandHandler {
   async execute(command: DeleteOrderedServiceCommand): Promise<void> {
     const orderedService = await this.orderedServiceRepo.findOne(command.orderedServiceId)
     if (!orderedService) throw new OrderedServiceNotFoundException()
+
+    // TODO: REFACTOR
+    const job = await this.prismaService.orderedJobs.findUnique({ where: { id: orderedService.getProps().jobId } })
+    if (!job) throw new JobNotFoundException()
+    const invoiceId = job.invoiceId
+
+    if (invoiceId !== null) {
+      const invoice = await this.prismaService.invoices.findUnique({ where: { id: invoiceId } })
+      if (!invoice) throw new InvoiceNotFoundException()
+      if (invoice.status !== 'Unissued') throw new IssuedJobUpdateException()
+    }
+
     await this.orderedServiceRepo.delete(orderedService.id)
     // TODO: History
   }
